refactor: change directory structure from redux-way to re-ducks
