+++ components/templates/Layout.tsx
@@ -2,7 +2,7 @@ import { createStyles, makeStyles, Theme } from "@material-ui/core/styles"
 import Head from "next/head"
 import * as React from "react"
 import { useSelector } from "react-redux"
-import { IInitialState } from "../../store/states"
+import { selectedPageSelector } from "../../store/page"
 import { ResponsiveDrawer } from "../organisms"
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -16,8 +16,6 @@ const useStyles = makeStyles((theme: Theme) =>
 interface IProps {
   children: React.ReactNode
 }
-
-const selectedPageSelector = (state: IInitialState) => state.page.selectedPage
 
 export const Layout = function(props: IProps) {
   const { children } = props
