+++ pages/redux.tsx
@@ -12,8 +12,8 @@ import { SpacingPaper } from "../components/atoms"
 import { HeaderArticleContainer } from "../components/organisms"
 import { Layout } from "../components/templates"
 import { Page } from "../constants"
-import { CounterActions, IPagePayload, PageActions } from "../store/actions"
-import { IInitialState } from "../store/states"
+import { CounterActions, countSelector } from "../store/counter"
+import { IPagePayload, PageActions } from "../store/page"
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -28,8 +28,6 @@ const useStyles = makeStyles((theme: Theme) =>
     },
   })
 )
-
-const countSelector = (state: IInitialState) => state.counter.count
 
 function Redux() {
   const classes = useStyles({})
