+++ store/counter/index.ts
@@ -1,0 +1,3 @@
+export * from "./actions"
+export * from "./selectors"
+export * from "./states"
