+++ store/counter/selectors.ts
@@ -1,0 +1,3 @@
+import { IInitialState } from "../states"
+
+export const countSelector = (state: IInitialState) => state.counter.count
