+++ store/counter/states.ts
@@ -1,0 +1,9 @@
+/**
+ * Counter
+ */
+export interface ICounterState {
+  count: number
+}
+export const CounterInitialState: ICounterState = {
+  count: 1,
+}
