+++ store/page/selectors.ts
@@ -1,0 +1,4 @@
+import { IInitialState } from "../states"
+
+export const selectedPageSelector = (state: IInitialState) =>
+  state.page.selectedPage
