+++ store/page/states.ts
@@ -1,0 +1,11 @@
+import { Page } from "../../constants"
+
+/**
+ * Page
+ */
+export interface IPageState {
+  selectedPage: Page
+}
+export const PageInitialState: IPageState = {
+  selectedPage: Page.TOP,
+}
