+++ store/reducers.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux"
-import { countReducers } from "./reducers/CounterReducer"
-import { pageReducers } from "./reducers/PageReducer"
+import countReducers from "./counter/reducers"
+import pageReducers from "./page/reducers"
 import { IInitialState } from "./states"
 
 export const combinedReducers = combineReducers<IInitialState>({
