--- store/reducers/index.ts
@@ -1,2 +1,0 @@
-export * from "./CounterReducer"
-export * from "./PageReducer"
